+++ lib/enums.dart
@@ -1,10 +1,3 @@
-enum CollectionType {
-  bukhari,
-  muslim,
-  nasai,
-  abudawud,
-  ibnmajah,
-  tirmidhi,
-}
+enum Collection { bukhari, muslim, nasai, abudawud, ibnmajah, tirmidhi }
 
 enum Languages { en, ar }