+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+- Reduced hadith loading time.
+- New classes & functions
+
 ## 1.0.0
 
 - Initial version.